+++ test/integration/cluster_env_test.go
@@ -28,7 +28,10 @@ import (
 )
 
 func TestClusterEnv(t *testing.T) {
-	minikubeRunner := util.MinikubeRunner{BinaryPath: *binaryPath, T: t}
+	minikubeRunner := util.MinikubeRunner{
+		Args:       *args,
+		BinaryPath: *binaryPath,
+		T:          t}
 	minikubeRunner.EnsureRunning()
 
 	dockerEnvVars := minikubeRunner.RunCommand("docker-env", true)
